+++ Reprise.SampleApi/Endpoints/Users/CreateUserEndpoint.cs
@@ -4,9 +4,8 @@
     public class CreateUserEndpoint
     {
         [Post("/users")]
-        public static IResult Handle(UserDto userDto, IValidator<UserDto> validator, DataContext context)
+        public static IResult Handle(UserDto userDto, DataContext context)
         {
-            validator.ValidateAndThrow(userDto);
             var user = new User()
             {
                 Id = !context.Users.Any() ? 0 : context.Users.Max(u => u.Id) + 1,