+++ Reprise.UnitTests/Features/ExceptionHandling/ExceptionHandlerTests.cs
@@ -11,10 +11,14 @@
 
         private readonly Mock<IErrorResponseFactory> _MockErrorResponseFactory = new();
 
-        private readonly ILoggerProvider _LoggerProvider = LoggerRecording.Start();
+        private readonly Mock<ILoggerFactory> _MockLoggerFactory = new();
+
+        private readonly ILogger _LoggerProvider = LoggerRecording.Start();
 
         public ExceptionHandlerTests()
         {
+            _MockLoggerFactory.Setup(m => m.CreateLogger(It.IsAny<string>()))
+                .Returns(_LoggerProvider);
             _MockHttpContext.Setup(m => m.Response)
                 .Returns(_MockHttpResponse.Object);
         }
@@ -24,7 +28,7 @@
         {
             var serviceProvider = CreateServiceProvider();
 
-            _ = new ExceptionHandler(_LoggerProvider, serviceProvider, _ => Task.CompletedTask);
+            _ = new ExceptionHandler(_MockLoggerFactory.Object, serviceProvider, _ => Task.CompletedTask);
         }
 
         [Fact]
@@ -32,7 +36,7 @@
         {
             var serviceProvider = CreateServiceProvider(skipExceptionLogger: true);
 
-            return Throws(() => new ExceptionHandler(_LoggerProvider, serviceProvider, _ => Task.CompletedTask))
+            return Throws(() => new ExceptionHandler(_MockLoggerFactory.Object, serviceProvider, _ => Task.CompletedTask))
                 .IgnoreStackTrace();
         }
 
@@ -41,7 +45,7 @@
         {
             var serviceProvider = CreateServiceProvider(skipErrorResponseFactory: true);
 
-            return Throws(() => new ExceptionHandler(_LoggerProvider, serviceProvider, _ => Task.CompletedTask))
+            return Throws(() => new ExceptionHandler(_MockLoggerFactory.Object, serviceProvider, _ => Task.CompletedTask))
                 .IgnoreStackTrace();
         }
 
@@ -56,7 +60,7 @@
 
                 return Task.CompletedTask;
             };
-            var exceptionHandler = new ExceptionHandler(_LoggerProvider, serviceProvider, next);
+            var exceptionHandler = new ExceptionHandler(_MockLoggerFactory.Object, serviceProvider, next);
 
             await exceptionHandler.InvokeAsync(_MockHttpContext.Object);
 
@@ -68,7 +72,7 @@
         {
             var serviceProvider = CreateServiceProvider();
             RequestDelegate next = _ => throw new BadHttpRequestException("Test message");
-            var exceptionHandler = new ExceptionHandler(_LoggerProvider, serviceProvider, next);
+            var exceptionHandler = new ExceptionHandler(_MockLoggerFactory.Object, serviceProvider, next);
 
             await exceptionHandler.InvokeAsync(_MockHttpContext.Object);
 
@@ -81,7 +85,7 @@
         {
             var serviceProvider = CreateServiceProvider();
             RequestDelegate next = _ => throw new ValidationException("Test message");
-            var exceptionHandler = new ExceptionHandler(_LoggerProvider, serviceProvider, next);
+            var exceptionHandler = new ExceptionHandler(_MockLoggerFactory.Object, serviceProvider, next);
 
             await exceptionHandler.InvokeAsync(_MockHttpContext.Object);
 
@@ -94,7 +98,7 @@
         {
             var serviceProvider = CreateServiceProvider();
             RequestDelegate next = _ => throw new Exception("Test message");
-            var exceptionHandler = new ExceptionHandler(_LoggerProvider, serviceProvider, next);
+            var exceptionHandler = new ExceptionHandler(_MockLoggerFactory.Object, serviceProvider, next);
 
             await exceptionHandler.InvokeAsync(_MockHttpContext.Object);
 
@@ -109,7 +113,7 @@
             RequestDelegate next = _ => throw new Exception("Test message");
             _MockHttpResponse.Setup(m => m.HasStarted)
                 .Returns(true);
-            var exceptionHandler = new ExceptionHandler(_LoggerProvider, serviceProvider, next);
+            var exceptionHandler = new ExceptionHandler(_MockLoggerFactory.Object, serviceProvider, next);
 
             await exceptionHandler.InvokeAsync(_MockHttpContext.Object);
 
@@ -124,7 +128,7 @@
             RequestDelegate next = _ => throw new Exception("Test message");
             _MockErrorResponseFactory.Setup(m => m.Create(It.IsAny<ErrorContext<Exception>>()))
                 .Throws(new NotImplementedException());
-            var exceptionHandler = new ExceptionHandler(_LoggerProvider, serviceProvider, next);
+            var exceptionHandler = new ExceptionHandler(_MockLoggerFactory.Object, serviceProvider, next);
 
             await exceptionHandler.InvokeAsync(_MockHttpContext.Object);