+++ Reprise/Features/ExceptionHandling/ExceptionHandler.cs
@@ -10,9 +10,9 @@
 
         private readonly RequestDelegate _Next;
 
-        public ExceptionHandler(ILoggerProvider loggerProvider, IServiceProvider serviceProvider, RequestDelegate next)
+        public ExceptionHandler(ILoggerFactory loggerFactory, IServiceProvider serviceProvider, RequestDelegate next)
         {
-            _Logger = loggerProvider.CreateLogger(GetType().FullName ?? nameof(ExceptionHandler));
+            _Logger = loggerFactory.CreateLogger(GetType().FullName ?? nameof(ExceptionHandler));
             _ExceptionLogger = serviceProvider.GetRequiredServiceSafe<IExceptionLogger>();
             _ErrorResponseFactory = serviceProvider.GetRequiredServiceSafe<IErrorResponseFactory>();
             _Next = next;
@@ -32,7 +32,7 @@
                 }
                 catch (Exception ex2)
                 {
-                    _Logger.LogError(ex2, "An exception was thrown attempting to execute the error handler.");
+                    _Logger.LogError(ex2, "An exception was thrown while executing the error handler.");
                     httpContext.Response.StatusCode = StatusCodes.Status500InternalServerError;
                 }
             }
@@ -63,14 +63,14 @@
                     response = _ErrorResponseFactory.Create(exceptionContext);
                     break;
             }
-            await WriteResponse(httpContext, statusCode, response);
+            await WriteResponse(exception, httpContext, statusCode, response);
         }
 
-        private async Task WriteResponse(HttpContext httpContext, int statusCode, object? response)
+        private async Task WriteResponse(Exception exception, HttpContext httpContext, int statusCode, object? response)
         {
             if (httpContext.Response.HasStarted)
             {
-                _Logger.LogError("The response has already started, the error handler will not be executed.");
+                _Logger.LogError(exception, "The response has already started, the error handler will not be executed.");
             }
             else
             {